Created new OutfitList component (spun off from ItemsList) to reduce complexity and redundancy of ItemsList, removed conditional rendering
